Moving around encoders and handing throwable returns
